refactor(pyproject.toml): update script entry point to base_app for better clarity refactor(cli): replace root app with base_app for improved module structure chore(cli): remove unused _types.py to clean up codebase
refactor(cli): move CLI logic from root.py to app.py for better organization and maintainability
